adding rest of example pages
