+++ examples/Examples.jsx
@@ -2,133 +2,10 @@
 import * as React from 'react';
 import './Examples.css';
 import './Examples_overrides.css';
-import classnames from 'classnames';
-
-import styled, { keyframes } from 'styled-components';
 import StarPicker from '../src/StarPicker';
-import type { StarRendererProps } from '../src/types';
 
-const LabelComp = ({
-  type,
-  label,
-}: {
-  type: 'right' | 'left',
-  label: ?string,
-}) => (
-  <div className={classnames('InputWithLabel-label', type)}>
-    {label && type === 'right' && <hr />}
-    {label}
-    {label && type === 'left' && <hr />}
-  </div>
-);
-
-const InputWithLabel = ({
-  label,
-  children,
-  placement = 'left',
-  className,
-}: {
-  label: string,
-  children: React.Node,
-  placement: 'left' | 'right',
-  className?: string,
-}) => {
-  return (
-    <div className={classnames('InputWithLabel', className, placement)}>
-      <LabelComp type="left" label={placement === 'left' && label} />
-      <div className="InputWithLabel-children">{children}</div>
-      <LabelComp type="right" label={placement === 'right' && label} />
-    </div>
-  );
-};
-
-const StyledInputWithLabel = styled(InputWithLabel)`
-  display: flex;
-  align-items: center;
-  color: transparent;
-  transition: color 1s;
-
-  .InputWithLabel-label {
-    flex-basis: 0;
-    flex-grow: 1;
-    display: flex;
-    align-items: center;
-
-    &.right {
-      margin-right: 1em;
-    }
-    &.left {
-      margin-left: 1em;
-    }
-
-    hr {
-      flex-grow: 1;
-      height: 1px;
-      border: 0;
-      background-color: transparent;
-      transition: background-color 1s;
-
-      margin: 0 5px;
-    }
-  }
-  :hover {
-    color: lightgray;
-
-    hr {
-      background-color: lightgray;
-    }
-  }
-`;
-
-const rotate = keyframes`
-  0% {
-    transform: rotate(0deg);
-  }
-
-  30% {
-    transform: rotate(10deg);
-  }
-
-  70% {
-    transform: rotate(-10deg);
-  }
-`;
-const DollarSvgContainer = styled.div`
-  display: flex;
-  &:hover {
-    animation: ${rotate} 0.3s ease-in-out;
-  }
-  padding: 0;
-  margin: 0;
-  font-size: 1.2rem;
-`;
-
-const dollarRenderer = ({
-  index,
-  hoverIndex,
-  selectedIndex,
-  size,
-}: StarRendererProps): React.Node => {
-  const selected = index <= selectedIndex;
-  const nextSelection = hoverIndex != null && index <= hoverIndex;
-  const hasHover = hoverIndex != null;
-
-  const colored = (selected && !hasHover) || (nextSelection && hasHover);
-  const color = colored ? '#1ca363' : 'grey';
-  const colorOutline = colored ? 'green' : 'grey';
-  return (
-    <DollarSvgContainer>
-      <svg style={{ padding: 5 }} width={size * 0.7} viewBox="0 0 15 22">
-        <path
-          d="M 5.7672275,0.10454656 5.7687798,1.8432524 C 3.1530349,2.1696584 0.85511786,3.5105694 0.63651923,5.9335327 0.51636012,10.566614 10.734409,11.654347 10.934687,15.030807 10.864893,17.460972 4.2894804,17.706665 3.2163668,14.995092 L 0.02261549,15.720285 C 0.77014071,18.029563 3.1584552,19.322457 5.7677453,19.684782 L 5.7687798,21.539435 8.6509212,21.539953 8.6504035,19.710663 C 11.644631,19.349577 14.241035,17.773814 14.225752,15.097063 14.229088,9.1564354 4.1341996,9.133676 3.9524306,6.0044467 3.9494677,3.7582878 10.035145,3.326126 10.629806,6.0577627 L 13.68535,5.4536939 C 13.196125,3.4596776 11.059157,2.2534069 8.6509212,1.8856981 L 8.6504035,0.10506428 Z"
-          stroke={colorOutline}
-          fill={selected ? color : 'none'}
-          strokeWidth="0.9"
-        />
-      </svg>
-    </DollarSvgContainer>
-  );
-};
+import StyledInputWithLabel from './ExamplesHelpers';
+import dollarRenderer from './customRenderer_dollar';
 
 type Props = {};
 type State = {
@@ -158,7 +35,10 @@ class Examples extends React.Component<Props, State> {
 
     return (
       <div className="Examples">
-        <StyledInputWithLabel label="Customizable color mixings">
+        <StyledInputWithLabel
+          label="Customizable color mixing"
+          linkRef="#/Examples/Color%20mixing"
+        >
           <StarPicker
             name="rating1"
             value={rating1}
@@ -173,7 +53,10 @@ class Examples extends React.Component<Props, State> {
           />
         </StyledInputWithLabel>
 
-        <StyledInputWithLabel label="Half-star support">
+        <StyledInputWithLabel
+          label="Half-star support"
+          linkRef="#/Examples/Half%20stars"
+        >
           <StarPicker
             name="rating2"
             value={rating2}
@@ -191,6 +74,7 @@ class Examples extends React.Component<Props, State> {
         <StyledInputWithLabel
           label="Custom star renderer with an SVG"
           placement="right"
+          linkRef="#/Examples/Using%20a%20Custom%20Renderer"
         >
           <StarPicker
             name="rating3"
@@ -204,6 +88,7 @@ class Examples extends React.Component<Props, State> {
         <StyledInputWithLabel
           label="Customizable star characters"
           placement="right"
+          linkRef="#/Examples/Changing%20character"
         >
           <StarPicker
             name="rating4"
@@ -219,7 +104,11 @@ class Examples extends React.Component<Props, State> {
           />
         </StyledInputWithLabel>
 
-        <StyledInputWithLabel label="Customizable style" placement="right">
+        <StyledInputWithLabel
+          label="Customizable style"
+          placement="right"
+          linkRef="#/Examples/Styling%20the%20Default%20Renderer"
+        >
           <StarPicker
             name="rating5"
             className="enlargeStar"
@@ -233,7 +122,7 @@ class Examples extends React.Component<Props, State> {
           />
         </StyledInputWithLabel>
 
-        <StyledInputWithLabel label="(disabled)">
+        <StyledInputWithLabel label="(disabled)" linkRef="#/Examples/Disabling">
           <StarPicker
             value={2}
             onChange={this.setValue}
